add translation units
